update: 监测浏览器功能由utils统一提供
