+++ packages/utils/src/index.ts
@@ -18,6 +18,8 @@ export * from './string';
 
 export * from './uid';
 
+export * from './os';
+
 export const toLine = (name: string = '') => name.replace(/\B([A-Z])/g, '-$1').toLowerCase();
 
 export function replacePropertyWithValue(obj: Record<string | number | symbol, any>, value: any, newValue: any) {
