+++ packages/vue-aide/src/types.ts
@@ -1,11 +1,4 @@
-export type OperatingSystem = 'macOS' | 'Windows' | 'Android' | 'iOS' | 'Linux' | 'BSD' | 'Unknown';
-
-export type BrowserInfo = 'Chrome' | 'Firefox' | 'Safari' | 'Edge' | 'Internet Explorer' | 'Unknown';
-
-export interface Geolocation {
-  latitude: number;
-  longitude: number;
-}
+import { BrowserInfo, Geolocation, OperatingSystem } from '@pictode/utils';
 
 export interface OSContext {
   OS: OperatingSystem;
